requirements.txt
cianparser
pandas
flask
matplotlib
seaborn
numpy
ipykernel
scikit-learn
dvc
pyyaml